+++ .eslintrc.json
@@ -7,7 +7,7 @@
   "rules": {
     "no-var": "error",
     "semi": "error",
-    "indent": "error",
+    "indent": ["error", 2],
     "no-multi-spaces": "error",
     "space-in-parens": "error",
     "no-multiple-empty-lines": "error",
