+++ src/classes/Crypto/Crypto.ts
@@ -6,11 +6,11 @@ export class Crypto implements ICtypro {
     id: number;
 
     constructor(id: number, ccyId: number, balance: number, numberSign: number) {
-        this.id = id;
-        this.ccyId = ccyId;
-        this.balance = Crypto.trimsBalance(balance, numberSign);
+      this.id = id;
+      this.ccyId = ccyId;
+      this.balance = Crypto.trimsBalance(balance, numberSign);
     }
     private static trimsBalance(balance, to): number {
-        return pruningNumbers(balance, to);
+      return pruningNumbers(balance, to);
     }
 }
