+++ src/classes/User/UserTrader.ts
@@ -10,44 +10,44 @@ export class UserTrader extends UserGoals implements IUserTrader {
     private orders: Array<IOrder> = ordersDb;
 
     protected static getOrderInfo(order: IOrder): { give: string; quantity: number; get: string; id: number; typeOrderId: string } {
-        return {
-            id: order.id,
-            typeOrderId: typeOrder[order.typeOrderId],
-            quantity: order.quantity,
-            give: typeCCY[order.give],
-            get: typeCCY[order.get],
-        };
+      return {
+        id: order.id,
+        typeOrderId: typeOrder[order.typeOrderId],
+        quantity: order.quantity,
+        give: typeCCY[order.give],
+        get: typeCCY[order.get],
+      };
     }
     protected static printOrders (arr: Array<IOrder>): string {
-        const ordersInfo = arr.map(item => UserTrader.getOrderInfo(item));
-        const result = ordersInfo.map(item => `Ордер- ${item.id}: ${item.typeOrderId} - ${item.quantity} монет. Отдаю - ${item.give}, получаю - ${item.get}`);
-        return result.join(', \t\n');
+      const ordersInfo = arr.map(item => UserTrader.getOrderInfo(item));
+      const result = ordersInfo.map(item => `Ордер- ${item.id}: ${item.typeOrderId} - ${item.quantity} монет. Отдаю - ${item.give}, получаю - ${item.get}`);
+      return result.join(', \t\n');
     }
     public getAllOrders(): Array<IOrder> {
-        return this.orders;
+      return this.orders;
     }
 
     public getOrderById(id: number): IOrder {
-        return <IOrder>getById(this.orders, id);
+      return <IOrder>getById(this.orders, id);
     }
 
     public printAllOrders(): string {
-        return UserTrader.printOrders(this.orders);
+      return UserTrader.printOrders(this.orders);
     }
 
     public printOrderById(id: number): string {
-        const order = getById(this.orders, id);
+      const order = getById(this.orders, id);
 
-        return UserTrader.printOrders([order]);
+      return UserTrader.printOrders([order]);
     }
 
     public removeOrderByKey(id: number): void {
-        this.orders = removeById(this.orders, id);
+      this.orders = removeById(this.orders, id);
     }
 
     public setOrder(id: number, typeOrderId: typeOrder, quantity: number, give: typeCCY, get: typeCCY): void {
-        const order = new Order(id, typeOrderId, quantity, give, get);
-        this.orders.push({...order});
+      const order = new Order(id, typeOrderId, quantity, give, get);
+      this.orders.push({...order});
     }
 
 }
