+++ src/db/ccyDb.ts
@@ -2,8 +2,8 @@ import {ICCY} from "../interfaces/CCY/ICCY";
 import {typeCCY} from "../interfaces/CCY/ICCY";
 
 export const ccyDb: Array<ICCY> = [
-    {id: typeCCY.PRIZM, name: 'PRIZM', shortName: 'PZM', numberSign: 2},
-    {id: typeCCY.Umi, name: 'Umi', shortName: 'UMI', numberSign: 4},
-    {id: typeCCY.Bitcoin, name: 'Bitcoin', shortName: 'BTC', numberSign: 8},
-    {id: typeCCY.Alti, name: 'Alti', shortName: 'ALT', numberSign: 1},
+  {id: typeCCY.PRIZM, name: 'PRIZM', shortName: 'PZM', numberSign: 2},
+  {id: typeCCY.Umi, name: 'Umi', shortName: 'UMI', numberSign: 4},
+  {id: typeCCY.Bitcoin, name: 'Bitcoin', shortName: 'BTC', numberSign: 8},
+  {id: typeCCY.Alti, name: 'Alti', shortName: 'ALT', numberSign: 1},
 ];
